+++ packages/drivers/odsp-driver/src/odspDocumentServiceFactoryCore.ts
@@ -31,9 +31,8 @@ import {
     ICacheAndTracker,
 } from "./epochTracker";
 import { OdspDocumentService } from "./odspDocumentService";
-import { INewFileInfo, getOdspResolvedUrl, createOdspLogger } from "./odspUtils";
+import { INewFileInfo, getOdspResolvedUrl, createOdspLogger, toInstrumentedOdspTokenFetcher } from "./odspUtils";
 import { createNewFluidFile } from "./createFile";
-import { toInstrumentedOdspTokenFetcher } from "./prefetchSnapshotUtils";
 
 /**
  * Factory for creating the sharepoint document service. Use this if you want to
