+++ packages/drivers/odsp-driver/src/prefetchLatestSnapshot.ts
@@ -20,13 +20,14 @@ import { ChildLogger, PerformanceEvent, TelemetryLogger } from "@fluidframework/
 import { ISnapshotCacheValue } from "./odspDocumentStorageManager";
 import {
     createOdspLogger,
+    evalBlobsAndTrees,
     fetchAndParseAsJSONHelper,
     getOdspResolvedUrl,
     getWithRetryForTokenRefresh,
     IOdspResponse,
+    toInstrumentedOdspTokenFetcher,
 } from "./odspUtils";
 import { IOdspSnapshot } from "./contracts";
-import { evalBlobsAndTrees, toInstrumentedOdspTokenFetcher } from "./prefetchSnapshotUtils";
 
 /**
  * Function to prefetch the snapshot and cached it in the persistant cache, so that when the container is loaded
